backend/requirements.txt
Flask
pymongo
pandas
scikit-learn
python-dotenv